requirements.txt
mysql-connector-python 
python-barcode
pyautogui